changed qr activity to binomial activity and create count activity for qr
