+++ app/src/main/AndroidManifest.xml
@@ -22,20 +22,20 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Trialio">
-        <activity android:name=".activities.ScanningActivity"
+        <activity android:name=".activities.QRCountActivity"></activity>
+        <activity
+            android:name=".activities.ScanningActivity"
             android:parentActivityName=".activities.ExperimentActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.example.trialio.activities.MainActivity" />
         </activity>
-        <activity android:name=".controllers.QRCodeGenerator"></activity>
-  
+        <activity android:name=".controllers.QRCodeGenerator" />
         <activity android:name=".activities.TrialActivity" />
         <activity android:name=".activities.ExperimentSettingsActivity" />
         <activity android:name=".activities.QuestionForumActivity" />
-        <activity android:name=".activities.QuestionRepliesActivity"></activity>
+        <activity android:name=".activities.QuestionRepliesActivity" />
         <activity android:name=".activities.ViewUserActivity" />
-
         <activity
             android:name=".activities.ExperimentActivity"
             android:parentActivityName=".activities.MainActivity">
@@ -72,7 +72,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.QRCodeActivity" />
+        <activity android:name=".activities.QRBinomialActivity" />
     </application>
 
 </manifest>
