+++ app/src/main/java/com/example/trialio/activities/QRBinomialActivity.java
@@ -1,0 +1,44 @@
+package com.example.trialio.activities;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.trialio.R;
+import com.example.trialio.controllers.ExperimentManager;
+import com.example.trialio.models.Experiment;
+import com.example.trialio.models.Trial;
+
+import java.util.ArrayList;
+
+import javax.annotation.Nullable;
+
+public class QRBinomialActivity extends AppCompatActivity {
+    private Experiment experiment;
+    private ArrayList<Trial> trialList;
+    private ExperimentManager experimentManager;
+
+
+    /**
+     * onCreate takes in the experiment passed in as a bundle and send triallist into QRAdaptor
+     * @param savedInstanceState
+     */
+    @Override
+    @Nullable
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_qr_binomial);
+
+        // get the experiment that was passed in
+        Bundle bundle = getIntent().getExtras();
+        experiment = (Experiment) bundle.getSerializable("experiment_qr");
+
+        trialList = experiment.getTrialManager().getTrials();
+
+        experimentManager = new ExperimentManager();
+    }
+
+
+
+
+}
