+++ app/src/main/java/com/example/trialio/activities/QRCountActivity.java
@@ -1,0 +1,16 @@
+package com.example.trialio.activities;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.trialio.R;
+
+public class QRCountActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_qr_count);
+    }
+}
