+++ app/src/main/res/layout/activity_qr_binomial.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#747474"
-    tools:context=".activities.QRCodeActivity">>
+    tools:context=".activities.QRBinomialActivity">>
 
 
     <LinearLayout
@@ -78,16 +78,6 @@
         android:layout_marginLeft="5dp"
         android:layout_marginRight="5dp"
         android:gravity="center">
-
-        <ListView
-            android:id="@+id/list_trials_QR"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:background="#9c9c9c"
-            android:focusable="true">
-
-        </ListView>
 
     </LinearLayout>
