+++ javadocs/com/example/trialio/activities/QRCodeActivity.html
@@ -123,7 +123,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.appcompat.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>com.example.trialio.activities.QRCodeActivity</li>
+<li>com.example.trialio.activities.QRBinomialActivity</li>
 </ul>
 </li>
 </ul>
